pyproject.toml
[tool.poetry]
name = "arara-quant-lab"
version = "0.1.0"
description = "Arara Quant Lab's multi-asset portfolio research toolkit for the PRISM-R strategy."
authors = ["Marcus Vinícius Silva", "Anna Beatriz Cardoso"]
readme = "README.md"
license = "MIT"
packages = [{include = "arara_quant", from = "src"}]

[tool.poetry.scripts]
arara-quant = "arara_quant.cli:main"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
pandas = "^2.2.2"
numpy = "^2.1.0"
cvxpy = "^1.4.3"
scikit-learn = "^1.4.2"
osqp = "^0.6.5"
clarabel = "^0.8.0"
yfinance = "^0.2.41"
pandas-datareader = "^0.10.0"
matplotlib = "^3.8.4"
typing-extensions = "^4.11.0"
pyarrow = "^21.0.0"
pydantic = "^2.0"
pyyaml = "^6.0"
jinja2 = "^3.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
ruff = "^0.4.2"
black = "^24.4.2"
mypy = "^1.9.0"
pre-commit = "^3.7.0"
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
exclude = ["scripts/"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "PL"]
ignore = [
    "E501",      # Line too long
    "E402",      # Module level import not at top of file
    "PLR2004",   # Magic value in comparison
    "PLR0911",   # Too many return statements
    "PLR0912",   # Too many branches
    "PLR0913",   # Too many arguments
    "PLR0915",   # Too many statements
    "PLW0603",   # Global statement
    "PLW2901",   # Loop variable overwritten
    "PLE0605",   # Invalid __all__ format
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = ["--import-mode=importlib"]

[tool.coverage.run]
source = ["src/arara_quant"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
    "src/arara_quant/__init__.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "def __str__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@overload",
]

[tool.coverage.html]
directory = "htmlcov"
